Relocate metrics tracker to plugin reporting
move createMetricsTracker from src/shared/utils/metrics.js into src/plugin/src/reporting/metrics.js so the plugin-only metrics helper lives alongside reporting code.

remove the shared re-export shim and stop exporting metrics from the shared utils barrel.
