--- src/shared/metrics-utils.js
@@ -1,1 +1,0 @@
-export * from "./utils/metrics.js";
